requirements.txt
absl-py==2.3.1
addict==2.4.0
aiofiles==25.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
appnope==0.1.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
arrow==1.4.0
asttokens==3.0.0
async-lru==2.0.5
async-timeout==5.0.1
attrs==25.4.0
av==16.1.0
babel==2.17.0
beautifulsoup4==4.14.2
black==25.11.0
bleach==6.3.0
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cloudpickle==3.1.2
colorama==0.4.6
coloredlogs==15.0.1
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
Cython==3.2.1
debugpy==1.8.17
decorator==5.2.1
deepgram-sdk==5.3.2
defusedxml==0.7.1
# Editable install with no version control (detectron2==0.6)
-e /Users/aloshdenny/vscode/TubeletGraph/thirdparty/detectron2
distro==1.9.0
docstring_parser==0.17.0
elevenlabs==2.35.0
eva-decord==0.6.1
eval_type_backport==0.3.1
exceptiongroup==1.3.0
executing==2.2.1
fastjsonschema==2.21.2
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.60.1
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2025.10.0
ftfy==6.3.1
fvcore==0.1.5.post20221221
gdown==5.2.0
googleapis-common-protos==1.72.0
groq==1.0.0
grpcio==1.76.0
h11==0.16.0
h5py==3.15.1
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.11
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.1
iopath==0.1.9
ipykernel==7.1.0
ipython==8.37.0
ipywidgets==8.1.8
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
json5==0.12.1
jsonpointer==3.0.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.3.0
jupyter_client==8.6.3
jupyter_core==5.9.1
jupyter_server==2.17.0
jupyter_server_terminals==0.5.3
jupyterlab==4.4.10
jupyterlab_pygments==0.3.0
jupyterlab_server==2.28.0
jupyterlab_widgets==3.0.16
kiwisolver==1.4.9
lark==1.3.1
lazy_loader==0.4
livekit==1.0.25
livekit-agents==1.4.1
livekit-api==1.1.0
livekit-blingfire==1.1.0
livekit-plugins-deepgram==1.4.1
livekit-plugins-elevenlabs==1.4.1
livekit-plugins-openai==1.4.1
livekit-plugins-silero==1.4.1
livekit-protocol==1.1.2
Markdown==3.10
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.1
matplotlib-inline==0.2.1
mdurl==0.1.2
mediapy==1.2.4
mistune==3.1.4
mmcv==2.2.0
mmengine==0.10.7
mpmath==1.3.0
multidict==6.7.1
mypy_extensions==1.1.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.4.7
notebook_shim==0.2.4
numpy==2.1.2
omegaconf==2.3.0
onnxruntime==1.23.2
open_clip_torch==3.2.0
openai==2.21.0
opencv-python==4.11.0.86
opencv-python-headless==4.12.0.88
opentelemetry-api==1.39.1
opentelemetry-exporter-otlp==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-semantic-conventions==0.60b1
orjson==3.11.7
overrides==7.7.0
packaging==24.2
pandas==2.3.3
pandocfilters==1.5.1
parso==0.8.5
pathspec==0.12.1
pexpect==4.9.0
pillow==12.0.0
pinecone==8.0.1
pinecone-plugin-assistant==3.0.2
pinecone-plugin-interface==0.0.7
platformdirs==4.5.0
portalocker==3.2.0
prometheus_client==0.23.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.1
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
PyAudio==0.2.14
pycocotools==2.0.8
pycparser==2.23
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
PyJWT==2.11.0
pyparsing==3.2.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.2.1
python-json-logger==4.0.0
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
reportlab==4.4.4
requests==2.32.5
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rfc3987-syntax==1.1.0
rich==14.2.0
rpds-py==0.28.0
safetensors==0.6.2
# Editable install with no version control (SAM-2==1.0)
-e /Users/aloshdenny/vscode/TubeletGraph/thirdparty/sam2
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
Send2Trash==1.8.3
sentence-transformers==5.2.2
shapely==2.1.2
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.5
soupsieve==2.8
stack-data==0.6.3
submitit==1.5.3
sympy==1.14.0
tabulate==0.9.0
tensorboard==2.20.0
tensorboard-data-server==0.7.2
termcolor==3.2.0
terminado==0.18.1
threadpoolctl==3.6.0
tifffile==2025.5.10
timm==1.0.22
tinycss2==1.4.0
tokenizers==0.22.2
tomli==2.3.0
torch==2.9.1
torchvision==0.24.1
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==5.1.0
typer==0.23.1
typer-slim==0.20.0
types-protobuf==6.32.1.20251210
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uri-template==1.3.0
urllib3==2.5.0
watchfiles==1.1.1
wcwidth==0.2.14
webcolors==25.10.0
webencodings==0.5.1
websocket-client==1.9.0
websockets==15.0.1
Werkzeug==3.1.3
widgetsnbextension==4.0.15
yacs==0.1.8
yapf==0.43.0
yarl==1.22.0
zipp==3.23.0
